+++ conf.py
@@ -17,6 +17,10 @@ products = [
                 'product': '2.2.5',
                 'python': '39',
             },
+            {
+                'product': '2.4.1',
+                'python': '39',
+            },
         ]
     },
     {
@@ -25,12 +29,16 @@ products = [
             {
                 'product': '0.23.0',
                 'authorizer': '0.2.0',
+            },
+            {
+                'product': '24.0.0',
+                'authorizer': '0.4.0',
             }
         ]
     },
     {
         'name': 'hadoop',
-        'versions': ['3.2.2', '3.3.1', '3.3.3'],
+        'versions': [{'product': '3.2.2'}, {'product': '3.3.1'}, {'product': '3.3.3'}, {'product': '3.3.4'}],
     },
     {
         'name': 'hbase',
@@ -63,6 +71,7 @@ products = [
             {
                 'product': '2.3.9',
                 'hadoop': '2.10.1',
+                'jackson_dataformat_xml': '2.7.9',
                 'aws_java_sdk_bundle': '1.11.271',
                 'azure_storage': '7.0.1',
                 'azure_keyvault_core': '1.0.0',
@@ -70,6 +79,7 @@ products = [
             {
                 'product': '3.1.3',
                 'hadoop': '3.3.3',
+                'jackson_dataformat_xml': '2.12.3',
                 'aws_java_sdk_bundle': '1.11.1026',
                 'azure_storage': '7.0.1',
                 'azure_keyvault_core': '1.0.0',
@@ -79,10 +89,6 @@ products = [
     {
         'name': 'java-base',
         'versions': [
-            {
-                'product': '1.8.0',
-                '_security_path': '/usr/lib/jvm/jre-1.8.0/lib/security/java.security',
-            },
             {
                 'product': '11',
                 '_security_path': '/usr/lib/jvm/jre-11/conf/security/java.security',
@@ -116,15 +122,25 @@ products = [
                 'scala': '2.13',
                 'opa_authorizer': '1.4.0',
             },
+            {
+                'product': '3.3.1',
+                'scala': '2.13',
+                'opa_authorizer': '1.4.0',
+            },
         ]
     },
     {
+        'name': 'vector',
+        'versions': ['0.26.0'],
+    },
+    {
         'name': 'nifi',
-        'versions': ['1.13.2', '1.15.0', '1.15.1', '1.15.2', '1.15.3', '1.16.0', '1.16.1', '1.16.2', '1.16.3'],
+        'versions': [{'product': '1.15.3'}, {'product': '1.16.3'}, {'product': '1.18.0'}],
     },
     {
         'name': 'opa',
-        'versions': ['0.27.1', '0.28.0', '0.37.2', '0.41.0'],
+        'versions': [{'product': '0.27.1'}, {'product': '0.28.0'}, {'product': '0.37.2'}, {'product': '0.41.0'},
+                     {'product': '0.45.0'}],
     },
     {
         'name': 'pyspark-k8s',
@@ -146,19 +162,6 @@ products = [
         ]
     },
     {
-        'name': 'spark',
-        'versions': [
-            {
-                'product': '3.0.1',
-                'hadoop': '2.7',
-            },
-            {
-                'product': '3.1.1',
-                'hadoop': '2.7',
-            },
-        ]
-    },
-    {
         'name': 'spark-k8s',
         'versions': [
             {
@@ -174,6 +177,10 @@ products = [
                 'azure_keyvault_core': '1.0.0',
             },
         ]
+    },
+    {
+        'name': 'stackable-base',
+        'versions': ['1.0.0'],
     },
     {
         'name': 'superset',
@@ -198,36 +205,46 @@ products = [
             {
                 'product': '377',
                 'java': '11',
+                'java_base_image_sha256': '7929833412c331fc23cde0e23ca730d652c0be61a8a69c8a82b2af937a3fbd4e',
                 'opa_authorizer': '0.1.0',
             },
             {
                 'product': '387',
                 'java': '11',
+                'java_base_image_sha256': '7929833412c331fc23cde0e23ca730d652c0be61a8a69c8a82b2af937a3fbd4e',
                 'opa_authorizer': '0.1.0'
             },
             {
                 'product': '395',
                 'java': '17',
+                'java_base_image_sha256': '2b8d60d1ab50d56240cb6286d6bc377410442afbfc3292d81be5674bc0b51724',
                 'opa_authorizer': 'stackable0.1.0'
             },
             {
                 'product': '396',
                 'java': '17',
+                'java_base_image_sha256': '2b8d60d1ab50d56240cb6286d6bc377410442afbfc3292d81be5674bc0b51724',
+                'opa_authorizer': 'stackable0.1.0'
+            },
+            {
+                'product': '403',
+                'java': '17',
+                'java_base_image_sha256': '2b8d60d1ab50d56240cb6286d6bc377410442afbfc3292d81be5674bc0b51724',
                 'opa_authorizer': 'stackable0.1.0'
             },
         ],
     },
     {
         'name': 'tools',
-        'versions': ['0.2.0'],
+        'versions': [{'product': '0.2.0'}],
     },
     {
         'name': 'testing-tools',
-        'versions': ['0.1.0'],
+        'versions': [{'product': '0.1.0'}],
     },
     {
         # ZooKeeper must be at least 3.5.0
         'name': 'zookeeper',
-        'versions': ['3.5.8', '3.6.3', '3.7.0', '3.8.0'],
+        'versions': [{'product': '3.5.8'}, {'product': '3.6.3'}, {'product': '3.7.0'}, {'product': '3.8.0'}],
     },
 ]