+++ airflow/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [airflow-stackable0.5.0] - 2022-11-30
+
+### Added
+
+- Add Trino dependency and reduce image size ([#264])
+
+[#264]: https://github.com/stackabletech/docker-images/pull/264
+
+## [airflow-stackable0.4.0] - 2022-10-12
+
+### Added
+
+- Support for Airflow 2.4.1 ([#215])
+
+[#215]: https://github.com/stackabletech/docker-images/pull/215
+
 ## [airflow-stackable0.3.0] - 2022-04-28
 
 ### Changed